FOR DEBUGGING ONLY - DO NOT USE: Add deleteSync flag for linking
Signal will not allow linking if any of your devices support
deletesyncing and the one you are trying to link doesn't, it will
throw a 409.

This commit is a work in progress to debug and overcome this issue
as I can't currently link my new device :-) With this change I can
link my account at the cost of breaking it's state.